Return ‘DeviceTree’ instead of ‘RestoreDeviceTree’ from get_component_name().
Returning ‘RestoreDeviceTree’ is inconsistent with other component names, and the caller
apparently handles the need for it. Also, the API doc for plist_dictionary_insert_item states
that it asserts if the key to be inserted already exists. if plist_dictionary_insert_item actually
does assert as described, the caller of get_component_name triggers that assertion if
‘RestoreDeviceTree’ is returned.